requirements.txt
Flask
passlib[bcrypt]
mysql-connector-python
python-dotenv
